Fix PKCS 1.5 padding issue and add framework for challenge 48
PKCS1.5 uses random bytes in padding for mode 02 blocks, not 0xff.
This makes the attack harder. Challenge 48 currently uses same
code as challenge 47. It sometimes works but is not reliable. Even
challenge 47 can fail a smaller percentage of the time.

Also added in a safeguard to make sure that RSA modulus is not fewer
bytes than specified bit length divided by 8.